+++ google/cloud/storage/constants.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Constants used acros google.cloud.storage modules."""
+"""Constants used across google.cloud.storage modules."""
 
 # Storage classes
 
@@ -119,13 +119,15 @@ See: https://cloud.google.com/storage/docs/public-access-prevention
 """
 
 RPO_ASYNC_TURBO = "ASYNC_TURBO"
-"""Turbo Replication RPO
+"""The recovery point objective (RPO) indicates how quickly newly written objects are asynchronously replicated to a separate geographic location.
+When the RPO value is set to ASYNC_TURBO, the turbo replication feature is enabled.
 
 See: https://cloud.google.com/storage/docs/managing-turbo-replication
 """
 
 RPO_DEFAULT = "DEFAULT"
-"""Default RPO
+"""The recovery point objective (RPO) indicates how quickly newly written objects are asynchronously replicated to a separate geographic location.
+When the RPO value is set to DEFAULT, the default replication behavior is enabled.
 
 See: https://cloud.google.com/storage/docs/managing-turbo-replication
 """