+++ tests/unit/test_bucket.py
@@ -230,6 +230,28 @@ class Test_LifecycleRuleConditions(unittest.TestCase):
         self.assertEqual(conditions.number_of_newer_versions, 3)
         self.assertEqual(conditions.noncurrent_time_before, noncurrent_before)
 
+    def test_ctor_w_matches_prefix(self):
+        conditions = self._make_one(matches_prefix=["test-prefix"])
+        expected = {"matchesPrefix": ["test-prefix"]}
+        self.assertEqual(dict(conditions), expected)
+        self.assertIsNone(conditions.age)
+        self.assertIsNone(conditions.created_before)
+        self.assertIsNone(conditions.is_live)
+        self.assertIsNone(conditions.matches_storage_class)
+        self.assertIsNone(conditions.matches_suffix)
+        self.assertEqual(conditions.matches_prefix, ["test-prefix"])
+
+    def test_ctor_w_matches_suffix(self):
+        conditions = self._make_one(matches_suffix=["test-suffix"])
+        expected = {"matchesSuffix": ["test-suffix"]}
+        self.assertEqual(dict(conditions), expected)
+        self.assertIsNone(conditions.age)
+        self.assertIsNone(conditions.created_before)
+        self.assertIsNone(conditions.is_live)
+        self.assertIsNone(conditions.matches_storage_class)
+        self.assertIsNone(conditions.matches_prefix)
+        self.assertEqual(conditions.matches_suffix, ["test-suffix"])
+
     def test_from_api_repr(self):
         import datetime
 
@@ -902,7 +924,7 @@ class Test_Bucket(unittest.TestCase):
     def test_path_w_name(self):
         NAME = "name"
         bucket = self._make_one(name=NAME)
-        self.assertEqual(bucket.path, "/b/%s" % NAME)
+        self.assertEqual(bucket.path, f"/b/{NAME}")
 
     def test_get_blob_miss_w_defaults(self):
         from google.cloud.exceptions import NotFound
@@ -918,7 +940,7 @@ class Test_Bucket(unittest.TestCase):
 
         self.assertIsNone(result)
 
-        expected_path = "/b/%s/o/%s" % (name, blob_name)
+        expected_path = f"/b/{name}/o/{blob_name}"
         expected_query_params = {"projection": "noAcl"}
         expected_headers = {}
         client._get_resource.assert_called_once_with(
@@ -952,7 +974,7 @@ class Test_Bucket(unittest.TestCase):
         self.assertIs(blob.bucket, bucket)
         self.assertEqual(blob.name, blob_name)
 
-        expected_path = "/b/%s/o/%s" % (name, blob_name)
+        expected_path = f"/b/{name}/o/{blob_name}"
         expected_query_params = {
             "userProject": user_project,
             "projection": "noAcl",
@@ -986,7 +1008,7 @@ class Test_Bucket(unittest.TestCase):
         self.assertEqual(blob.name, blob_name)
         self.assertEqual(blob.generation, generation)
 
-        expected_path = "/b/%s/o/%s" % (name, blob_name)
+        expected_path = f"/b/{name}/o/{blob_name}"
         expected_query_params = {
             "generation": generation,
             "projection": "noAcl",
@@ -1020,7 +1042,7 @@ class Test_Bucket(unittest.TestCase):
         self.assertEqual(blob.name, blob_name)
         self.assertEqual(blob.etag, etag)
 
-        expected_path = "/b/%s/o/%s" % (name, blob_name)
+        expected_path = f"/b/{name}/o/{blob_name}"
         expected_query_params = {
             "projection": "noAcl",
         }
@@ -1055,7 +1077,7 @@ class Test_Bucket(unittest.TestCase):
         self.assertEqual(blob.name, blob_name)
         self.assertEqual(blob.generation, generation)
 
-        expected_path = "/b/%s/o/%s" % (name, blob_name)
+        expected_path = f"/b/{name}/o/{blob_name}"
         expected_query_params = {
             "ifGenerationMatch": generation,
             "projection": "noAcl",
@@ -1093,7 +1115,7 @@ class Test_Bucket(unittest.TestCase):
         self.assertEqual(blob.chunk_size, chunk_size)
         self.assertEqual(blob._encryption_key, key)
 
-        expected_path = "/b/%s/o/%s" % (name, blob_name)
+        expected_path = f"/b/{name}/o/{blob_name}"
         expected_query_params = {
             "projection": "noAcl",
         }
@@ -1218,7 +1240,7 @@ class Test_Bucket(unittest.TestCase):
         self.assertIs(iterator, client._list_resource.return_value)
         self.assertIs(iterator.bucket, bucket)
 
-        expected_path = "/b/{}/notificationConfigs".format(bucket_name)
+        expected_path = f"/b/{bucket_name}/notificationConfigs"
         expected_item_to_value = _item_to_notification
         client._list_resource.assert_called_once_with(
             expected_path,
@@ -1246,7 +1268,7 @@ class Test_Bucket(unittest.TestCase):
         self.assertIs(iterator, other_client._list_resource.return_value)
         self.assertIs(iterator.bucket, bucket)
 
-        expected_path = "/b/{}/notificationConfigs".format(bucket_name)
+        expected_path = f"/b/{bucket_name}/notificationConfigs"
         expected_item_to_value = _item_to_notification
         other_client._list_resource.assert_called_once_with(
             expected_path,
@@ -1270,7 +1292,7 @@ class Test_Bucket(unittest.TestCase):
         with self.assertRaises(NotFound):
             bucket.get_notification(notification_id=notification_id)
 
-        expected_path = "/b/{}/notificationConfigs/{}".format(name, notification_id)
+        expected_path = f"/b/{name}/notificationConfigs/{notification_id}"
         expected_query_params = {}
         client._get_resource.assert_called_once_with(
             expected_path,
@@ -1319,7 +1341,7 @@ class Test_Bucket(unittest.TestCase):
         self.assertIsNone(notification.blob_name_prefix)
         self.assertEqual(notification.payload_format, JSON_API_V1_PAYLOAD_FORMAT)
 
-        expected_path = "/b/{}/notificationConfigs/{}".format(name, notification_id)
+        expected_path = f"/b/{name}/notificationConfigs/{notification_id}"
         expected_query_params = {"userProject": user_project}
         client._get_resource.assert_called_once_with(
             expected_path,
@@ -1519,7 +1541,7 @@ class Test_Bucket(unittest.TestCase):
         with self.assertRaises(NotFound):
             bucket.delete_blob(blob_name)
 
-        expected_path = "/b/%s/o/%s" % (name, blob_name)
+        expected_path = f"/b/{name}/o/{blob_name}"
         expected_query_params = {}
         client._delete_resource.assert_called_once_with(
             expected_path,
@@ -1542,7 +1564,7 @@ class Test_Bucket(unittest.TestCase):
 
         self.assertIsNone(result)
 
-        expected_path = "/b/%s/o/%s" % (name, blob_name)
+        expected_path = f"/b/{name}/o/{blob_name}"
         expected_query_params = {"userProject": user_project}
         client._delete_resource.assert_called_once_with(
             expected_path,
@@ -1565,7 +1587,7 @@ class Test_Bucket(unittest.TestCase):
 
         self.assertIsNone(result)
 
-        expected_path = "/b/%s/o/%s" % (name, blob_name)
+        expected_path = f"/b/{name}/o/{blob_name}"
         expected_query_params = {"generation": generation}
         client._delete_resource.assert_called_once_with(
             expected_path,
@@ -1592,7 +1614,7 @@ class Test_Bucket(unittest.TestCase):
 
         self.assertIsNone(result)
 
-        expected_path = "/b/%s/o/%s" % (name, blob_name)
+        expected_path = f"/b/{name}/o/{blob_name}"
         expected_query_params = {
             "ifGenerationMatch": generation,
             "ifMetagenerationMatch": metageneration,
@@ -1811,7 +1833,7 @@ class Test_Bucket(unittest.TestCase):
 
         bucket.reload(if_etag_match=etag)
 
-        expected_path = "/b/%s" % (name,)
+        expected_path = f"/b/{name}"
         expected_query_params = {
             "projection": "noAcl",
         }
@@ -1837,7 +1859,7 @@ class Test_Bucket(unittest.TestCase):
 
         bucket.reload(if_metageneration_match=metageneration_number)
 
-        expected_path = "/b/%s" % (name,)
+        expected_path = f"/b/{name}"
         expected_query_params = {
             "projection": "noAcl",
             "ifMetagenerationMatch": metageneration_number,
@@ -1899,7 +1921,7 @@ class Test_Bucket(unittest.TestCase):
 
         blob = mock.create_autospec(Blob)
         blob.name = blob_name
-        blob.path = "/b/{}/o/{}".format(bucket_name, blob_name)
+        blob.path = f"/b/{bucket_name}/o/{blob_name}"
         return blob
 
     def test_copy_blobs_wo_name(self):
@@ -2048,7 +2070,7 @@ class Test_Bucket(unittest.TestCase):
             _target_object=new_blob,
         )
 
-        expected_patch_path = "/b/{}/o/{}".format(dest_name, new_name)
+        expected_patch_path = f"/b/{dest_name}/o/{new_name}"
         expected_patch_data = {"acl": []}
         expected_patch_query_params = {"projection": "full"}
         client._patch_resource.assert_called_once_with(
@@ -2960,7 +2982,7 @@ class Test_Bucket(unittest.TestCase):
         from google.api_core.iam import Policy
 
         bucket_name = "name"
-        path = "/b/%s" % (bucket_name,)
+        path = f"/b/{bucket_name}"
         etag = "DEADBEEF"
         version = 1
         owner1 = "user:phred@example.com"
@@ -2994,7 +3016,7 @@ class Test_Bucket(unittest.TestCase):
         self.assertEqual(policy.version, api_response["version"])
         self.assertEqual(dict(policy), expected_policy)
 
-        expected_path = "/b/%s/iam" % (bucket_name,)
+        expected_path = f"/b/{bucket_name}/iam"
         expected_query_params = {}
         client._get_resource.assert_called_once_with(
             expected_path,
@@ -3010,7 +3032,7 @@ class Test_Bucket(unittest.TestCase):
         bucket_name = "name"
         timeout = 42
         user_project = "user-project-123"
-        path = "/b/%s" % (bucket_name,)
+        path = f"/b/{bucket_name}"
         etag = "DEADBEEF"
         version = 1
         api_response = {
@@ -3033,7 +3055,7 @@ class Test_Bucket(unittest.TestCase):
         self.assertEqual(policy.version, api_response["version"])
         self.assertEqual(dict(policy), expected_policy)
 
-        expected_path = "/b/%s/iam" % (bucket_name,)
+        expected_path = f"/b/{bucket_name}/iam"
         expected_query_params = {"userProject": user_project}
         client._get_resource.assert_called_once_with(
             expected_path,
@@ -3047,7 +3069,7 @@ class Test_Bucket(unittest.TestCase):
         from google.cloud.storage.iam import STORAGE_OWNER_ROLE
 
         bucket_name = "name"
-        path = "/b/%s" % (bucket_name,)
+        path = f"/b/{bucket_name}"
         etag = "DEADBEEF"
         version = 3
         owner1 = "user:phred@example.com"
@@ -3067,7 +3089,7 @@ class Test_Bucket(unittest.TestCase):
 
         self.assertEqual(policy.version, version)
 
-        expected_path = "/b/%s/iam" % (bucket_name,)
+        expected_path = f"/b/{bucket_name}/iam"
         expected_query_params = {"optionsRequestedPolicyVersion": version}
         client._get_resource.assert_called_once_with(
             expected_path,
@@ -3113,7 +3135,7 @@ class Test_Bucket(unittest.TestCase):
         self.assertEqual(returned.version, version)
         self.assertEqual(dict(returned), dict(policy))
 
-        expected_path = "%s/iam" % (bucket.path,)
+        expected_path = f"{bucket.path}/iam"
         expected_data = {
             "resourceId": bucket.path,
             "bindings": mock.ANY,
@@ -3177,7 +3199,7 @@ class Test_Bucket(unittest.TestCase):
         self.assertEqual(returned.version, version)
         self.assertEqual(dict(returned), dict(policy))
 
-        expected_path = "%s/iam" % (bucket.path,)
+        expected_path = f"{bucket.path}/iam"
         expected_data = {
             "resourceId": bucket.path,
             "bindings": mock.ANY,
@@ -3221,7 +3243,7 @@ class Test_Bucket(unittest.TestCase):
 
         self.assertEqual(found, expected)
 
-        expected_path = "/b/%s/iam/testPermissions" % (name,)
+        expected_path = f"/b/{name}/iam/testPermissions"
         expected_query_params = {}
         expected_query_params = {"permissions": permissions}
         client._get_resource.assert_called_once_with(
@@ -3256,7 +3278,7 @@ class Test_Bucket(unittest.TestCase):
 
         self.assertEqual(found, expected)
 
-        expected_path = "/b/%s/iam/testPermissions" % (name,)
+        expected_path = f"/b/{name}/iam/testPermissions"
         expected_query_params = {
             "permissions": permissions,
             "userProject": user_project,
@@ -3369,7 +3391,7 @@ class Test_Bucket(unittest.TestCase):
         )
 
         if not default_object_acl_loaded:
-            expected_path = "/b/%s/defaultObjectAcl" % (name,)
+            expected_path = f"/b/{name}/defaultObjectAcl"
             expected_query_params = {}
             client._get_resource.assert_called_once_with(
                 expected_path,
@@ -3581,7 +3603,7 @@ class Test_Bucket(unittest.TestCase):
         )
 
         if not default_object_acl_loaded:
-            expected_path = "/b/%s/defaultObjectAcl" % (name,)
+            expected_path = f"/b/{name}/defaultObjectAcl"
             expected_query_params = {}
             client._get_resource.assert_called_once_with(
                 expected_path,
@@ -3724,9 +3746,7 @@ class Test_Bucket(unittest.TestCase):
                     break
             else:  # pragma: NO COVER
                 self.fail(
-                    "Condition {} not found in {}".format(
-                        expected_condition, policy_conditions
-                    )
+                    f"Condition {expected_condition} not found in {policy_conditions}"
                 )
 
         return policy_fields, policy
@@ -3831,7 +3851,7 @@ class Test_Bucket(unittest.TestCase):
 
         bucket.lock_retention_policy(timeout=timeout, retry=retry)
 
-        expected_path = "/b/{}/lockRetentionPolicy".format(name)
+        expected_path = f"/b/{name}/lockRetentionPolicy"
         expected_data = None
         expected_query_params = {"ifMetagenerationMatch": metageneration}
         client._post_resource.assert_called_once_with(
@@ -3869,7 +3889,7 @@ class Test_Bucket(unittest.TestCase):
 
         bucket.lock_retention_policy()
 
-        expected_path = "/b/{}/lockRetentionPolicy".format(name)
+        expected_path = f"/b/{name}/lockRetentionPolicy"
         expected_data = None
         expected_query_params = {
             "ifMetagenerationMatch": metageneration,
@@ -3964,7 +3984,7 @@ class Test_Bucket(unittest.TestCase):
             )
         else:
             expected_api_access_endpoint = api_access_endpoint
-            expected_resource = "/{}".format(parse.quote(bucket_name))
+            expected_resource = f"/{parse.quote(bucket_name)}"
 
         if virtual_hosted_style or bucket_bound_hostname:
             expected_resource = "/"